MB-50537: Don't allow strict n2n encryption if external unencrypted dist
When attempting to enable node to node encryption (and setting the level
of it) we perform the following steps:

1) Enable external TLS listener
2) Setup net config
3) Disable (unused) external TCP listener

We can then change the encryption level by setting the security
settings:

4) Set cluster-encryption-level

At step 3 we may fail to tear down the external TCP listener for some
reason and it will live on. The couchbase-cli command invoking this will
fail, but the cluster reports that node to node encryption is enabled as
each node has set the appropriate config. If we then attempt to set the
cluster encryption level (4) to strict the command succeeds as all nodes
have encryption configured. Strict node to node encryption asserts that
we do not listen to any external TCP ports, and if we have not torn down
an external TCP listener then we are in violation of this.

Prevent the enabling of the "strict" cluster_encryption when an external
unencrypted distribution exists to prevent a violation of this by
checking for configured unencrypted distributions when we attempt to set
it.

Change-Id: Ie1f7f231d0042e851414100e6203253895e84447
Reviewed-on: https://review.couchbase.org/c/ns_server/+/179242
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>