Update menelaus_roles tests to use latest version roles
They currently use the 7.0 roles. If we use the latest version instead
of a specific one then we shouldn't have to update this for Elixir.Next.
We also need to mock cluster_compat_mode:is_developer_preview to prevent
an ETS lookup that will fail when we produce roles.

Change-Id: I2578e630cc1ad7a37b040337263af5323220da69
Reviewed-on: https://review.couchbase.org/c/ns_server/+/179690
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>