MB-53445, MB-51223: Error when stream end seen for stream being set up
The DCP proxy (dcp_consumer_conn in particular) handles replication set
up by sending a DCP_ADD_STREAM to the DCP consumer which issues a
DCP_STREAM_REQUEST to the DCP producer. The DCP proxy sets up streams as
instructed by the dcp_replicator but it will not reply to the
dcp_replicator until all streams have been set up (successfully or not).
It is possible for the producer of a replication connection to send a
stream end to the consumer before the proxy has processed the add stream
response. dcp_consumer_conn ignores such a scenario as it expects that
an add stream response (successful or not) will be made. There are
(error) cases in kv_engine which can cause the DCP consumer to fail to
send an add stream response and in such a scenario dcp_consumer_conn
will never reponsd to the dcp_replicator. This results in rebalance
hanging indefinitely. Another scenario also exists in which the add
stream response does arrive, but the stream has already been closed by
the producer and so no progress is made with the rebalance.

To address this issue, add a new handler for producer_stream_end
messages to dcp_consumer_conn which matches only when the
state == stream_state (we are setting up streams) that crashes the
dcp_proxy such that the rebalance will fail and the user can retry it.

Change-Id: I93165dc686dee79515495bce73c8128808779f62
Reviewed-on: https://review.couchbase.org/c/ns_server/+/179952
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>