Transition couchdb config profile key to inverse
Due to issues regarding proplists:get_bool and subtle logical bugs
that can be introduced by this, it makes more sense to generally only
add items to other profiles (not default) and should have a value of
"true" in the other profile s/t if it's missing, it'll fallback to
false. This may seem silly or pointless but this will allow us to skip
putting the inverse key in the default profile for every key in
another profile. Otherwise you may get strange results if the file is
not found for the default profile, since we will always fallback to an
empty profile if files cannot be found.

Change-Id: Ifcd6a4ff0e9f4044541d5d4d2431697b253f79a4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/180209
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>