Convert any default config profile entries
Removed all entries (besides 'name') and moved them to the other
profiles (just serverless, now) as the inverse of what they were in
the default profile (assuming they are boolean, which they were).

This leads to a problem where the implied return value of
proplists:get_bool is false if it's not found, and this can be
overlooked, in addition to the fact that if the default profile file
isn't found, we use a stand-in that's just a list with a single
proplist key of 'name'. If we allowed adding keys to the default
config, in addition to other configs, we would be forced to also
update any place where we use the empty default in the code. We also
don't want to remove the empty default, because it makes the code much
more resilient in the default case, which seems like a valid
trade-off.

In the end, it's much easier to modify all other profiles than it is
to modify all profiles, on every key addition. If we allow keys in the
default profile, we would have to start adding the inverses to every
key.

Change-Id: Ie73cade2b7f8c2ff7e9eb3a841b7707064fe21f5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/180210
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>