Add inverse key temporarily to fix mismatch
Configuration profile keys that are boolean require some special care
because of the default value for proplists:get_bool is false, even if
the key wasn't there at all. That means if you have a key in the
serverless profile that is false, you will need to put the true
variant in the other profiles (in this case, default) or you will have
suprising behavior. This can cause problems depending how the value is
used aftewards, but causes the couchdb http server not to start.

Change-Id: I6e9f7742200e076595e29b85cd2adffcc6080d47
Reviewed-on: https://review.couchbase.org/c/ns_server/+/180266
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>