MB-53445: Clarify that dcp_consumer_conn:stream_state stores opaques
We store a specific type of opaque in the fields to_add, to_close, and
to_close_on_producer of the stream_state in dcp_consumer_conn. This
opaque is provided by ns_server and memcached stores it in the
DcpConsumer object and is returned to ns_server in key responses such
as a DCP Add Stream response. When ns_server sets up the stream it
currently just uses the vBucket ID as the opaque; that is fine, but it
is non-obvious to readers that the elements in these fields are the
supplied opaque rather than the vBucket ID. This then introduces
confusion when we explicitly search for the opaques of packets in these
lists.

Make it clearar that these fields are not vBucket IDs by changing the
allowed type, and using the name Opaque throughout for them rather than
Parition.

Change-Id: I33d5137ca0e0960d584ec07b737dfa4369adc65c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/180442
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>