MB-53848 do the final config upgrade not when the config version
...reaches the current node version, but when it reaches the desired
version for the current upgrade.

For example 7.1 node might try to upgrade 6.6 cluster to 7.0. With
the current code rbac_upgrade key will not be deleted.

We still can upgrade the audit descriptors only when the cluster
version matches the node version, because the node doesn't know
the audit descriptors for the previous version of the cluster.

Change-Id: Ied8336bae9a96e34d02330aa4d3a423c0d7510f7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/180614
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>