MB-53932 Allowing loading sample bucket into existing bucket
This change allows loading of one or more pre-existing buckets with
sample data. The reason is to allow the serverless control plane to
create buckets with characteristics (e.g. magma storage backend) that
are different from the current behavior. While the current method
continues to work the new method is invoked by specifying a list of json
objects each of which is the name of the sample to use and the name of
the bucket into which to load the sample data.

In this example the travel-sample data is loaded into three different
buckets (which are required to be pre-created as the primary use case is
for serverless and we don't want any auto-creation done by loading
sample data to use the wrong type of bucket).

xcurl localhost:9000/sampleBuckets/install \
        -H "Content-Type: application/json" -d '[
    {
      "sample": "travel-sample",
      "bucket": "student1"
    },
    {
      "sample": "travel-sample",
      "bucket": "student2"
    },
    {
      "sample": "travel-sample",
      "bucket": "student3"
    }
  ]'

Change-Id: I28ab844a4e0f6f86ccaf51d575d0b2627ea98974
Reviewed-on: https://review.couchbase.org/c/ns_server/+/180869
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>