MB-53597: Collect tasks stats for all memcached taskables
Memcached now has a NoBucket taskable which has an increasing number of
tasks. To collect stats for this taskable (as well as the taskables of
all other Buckets) a new stats API (tasks-all) has been introduced. Use
it instead of the old "tasks" API.

Change-Id: Ifde5a5a45a6ccb29642a9fc8c0f2a5b69299c6f4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/181050
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>