MB-52623: Update revrpc url in case of password change
cbauth change:
https://review.couchbase.org/c/cbauth/+/181327

Each revrpc connection now has revrpc.UpdateURL() method which
is supposed to be called everytime when revrpc URL is changed.
Since the internal password is part of that URL, ns_server should
call that method each time the password is changed.

We don't want to drop revrpc connections on every change of password
(because making it seamless would require too many ns_server changes)
so in order to verify the connectivity UpdateURL sends an http request
for "<NewURL>/test" and expects it to return 200.

Change-Id: I141b4f8570997b6d39dfb52e4b31b6ef2dcb170d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/181331
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>