MB-55154: [BP] Improve node certificate validation
When a node certificate is about to be uploaded, we need to check if the
node name is mentioned in the SAN part of the certificate. If the node
is initialized and is already part of a cluster with other nodes, then
the node name must be mentioned, Otherwise a warning is generated.

The process of checking SAN part of a certificate must be rechecked
when a node is being added to a cluster. That means if the
new node/cluster node certificate has invalid SAN, the joining process
should fail. On the new node invalid cert will always fail,
and on a cluster node the process will fail only if cluster has only
one node.

Change-Id: I9d17fb86167839fcc1c039512c0bee175880b7eb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/181334
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/ns_server/+/185597
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>