MB-55206 Reapplying the SAN improvement code
After the bug found in serverless mode is resolved, we can re-merge
the changes made for cert SAN check improvements.

Original work was done here:
MB-53537: Improve node certificate validation

When a node certificate is about to be uploaded, we need to check if the
node name is mentioned in the SAN part of the certificate. If the node
is initialized and is already part of a cluster with other nodes, then
the node name must be mentioned, Otherwise a warning is generated.

The process of checking SAN part of a certificate must be rechecked
when a node is being added to a cluster. That means if the
new node/cluster node certificate has invalid SAN, the joining process
must fail.

Change-Id: I9512d95a664ba6c0d2cef3e5e9bb06c0b862a25b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/185578
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>