MB-53930: Introduce ingress_disabled return value
Introduce a constant and handler for a new return value to the REST
API. The constant is LIMIT_EXCEEDED which is the hex value of 0x35. To
replicate this you need to enable:

 1. Configure internalSettings/dataStorageLimit to 1
 2. Add data to a bucket until the storage exceeds 1 GiB
 3. Attempt to use ns_server REST API to insert into the bucket
 4. Should now work as expected

The ticket mostly refers to upsert but we don't actually do any
upserts ever from the REST api calls. We do simple 'set' requests
instead.

Change-Id: I95c754298b387298f6282072c594bfe1b16ab9f2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/182351
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>