MB-54133: fix secrets leak in REST api crash
Fixes a leak of basic auth user/password tuple when we encounter an
error in certain REST api requests. The ones identified for this
related to the addition of a server to a group. There may be other
places that are leaked in a similar fashion but I was unable to find
any specific places.

Change-Id: I4fdcc60ff1ec455a6a74a83d086a78f40985e897
Reviewed-on: https://review.couchbase.org/c/ns_server/+/181481
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>