MB-53123: Enable autofailover popups for broken 7.0.0 clusters
To address an issue in which we failed to add auto failover popup alert
config on upgrade from 6.X -> 7.0 for customers who are already on
7.0.X, re-enable the auto failover alerting on initialization of
menelaus_web_alerts_srv. To prevent enabling the popup alerts whenever
we restart a 7.0.5 node or join a 7.0.5 node to an existing 7.0.X
cluster we set a an extra config value,
popup_alerts_auto_failover_upgrade_70_fixed. We should be able to
remove this config value in upgrades to 7.1.X as the 7.0.X to 7.1.X
upgrade process re-enables auto failover popups.

Change-Id: I519c6ef398cc8f11cf30dfbd40c37730f2b1724b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/181665
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>