MB-53121: Send queued alert emails
When we enable alerts right now, it is not currently sending emails
for any of the already queued alerts. To fix this, we setup a pubsub
on the ns_config_events callback for the key in question --
"email_alerts". When this callback is triggered, we flush the existing
alerts and mark them as "sent". This doesn't ensure that we sent the
email, though. Instead it just makes sure we attempt to send the
email, at least. It didn't seem like we kept track of whether or not
we have succeeded in sending the emails, and never retry in the
existing code, so we kept the behaviour the same.

Change-Id: I0a483353b492912ab38033c3b7dbdf8e288b8df0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/181685
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>