MB-55030: Remove unnecessary "hack" config key
To correct a bug which failed to enable auto failover pop up alerts
during an upgrade we enabled them on 7.0.X clusters if a new config
key was not present. We then added the config key such that we did
not repeat the enabling of the pop up alerts. Upgrade to 7.1.X
already enabled pop up alerts via conventional means so add to the
upgrade path for 7.2.X to clean up the config key.

Change-Id: I4e64f52bf123380cd3caf710fd5ccac734b0d260
Reviewed-on: https://review.couchbase.org/c/ns_server/+/184972
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>