MB-54441 Fix broken ram quota restriction in bucket rest api
When a bucket has its ram quota updated immediately after creation of
the bucket, the bucket has a config entry but no servers added to it
yet, which causes the validation to fail, allowing arbritrarily large
ram quota.
The fix is simply to check for an empty server list, and resort to the
kv_nodes list instead.

In serverless, there is a very specific situation where a similar issue
occurs with durabilityMinLevel being able to be set to majority for a
bucket on 1 node. This is fixed with a check to the desired_servers
property of the bucket. While this bug occurs only with 1 node buckets,
so it should not occur in production, the fix may avoid similar bugs in
future.

Change-Id: I57901dae0b164e01224778f67e365327a9cfc1b8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/182435
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>