MB-54250: Remove support for views in bucket capabilities
Remove the 'couchapi' support from the bucket capabilities list that
can be seen by making requests to: /pools/default/buckets/* where it
produces a json like:

  "bucketCapabilities": [
    "collections",
    "durableWrite",
    "tombstonedUserXAttrs",
    "couchapi", <-- remove this
    "subdoc.ReplaceBodyWithXattr",
    "subdoc.DocumentMacroSupport",
    "subdoc.ReviveDocument",
    "preserveExpiry",
    "rangeScan",
    "dcp",
    "cbhello",
    "touch",
    "cccp",
    "xdcrCheckpointing",
    "nodesExt",
    "xattr"
  ],

This change has an affect on more than just this list, since we
changed everywhere that checks for couchapi capability. Ideally, this
should just work.

We also went ahead and removed a few other heuristics that are used by
the SDK to bootstrap / function. Another field that was modified was
the `http(s)CAPI` field(s) that included the port of the couchdb
endpoint. Additionally, we cleared out (set to empty string) the
`couchApiBase(HTTPS)` field(s).

Change-Id: I3d893cdea500870067a36b6eba55582f5f1cda4f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/182792
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>