MB-54625 Alert when mutation history size is approached
Once the mutation history size for a vbucket reaches the
history_warning_threshold (90% by default, but we could make this
configurable), an alert is given which warns that history retention
time may not be met and gives the number of vbuckets crossing the
warning threshold.
We check for vbuckets crossing the threshold by making a memcached
STAT request, instead of adding a stat for every vbucket to prometheus.

The alert only fires for a bucket once per 10mins (determined by
?ALERT_TIMEOUT). This is enforced by checking whether any other nodes
have fired the alert in #state.history for that bucket, and not
checking for additional alerts if they have.
The alternative approach to ensuring that only 1 alert fires per bucket
would have been to only check from the orchestrator/master node. This
would be more complex to implement, as it requires fetching vbucket
stats from all other nodes. It would either have to fetch these stats
every time it checks for the alert, or store the stats somewhere.

Example alert:
Warning: On bucket default mutation history is greater than 90% of
history retention size for 6/1024 vbuckets. Please ensure that
the history retention size is sufficiently large, in order for the
mutation history to be retained for the history retention time.

Change-Id: If7f013377ea8c4abd5623a7c12937e84e56dbd54
Reviewed-on: https://review.couchbase.org/c/ns_server/+/183189
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>