MB-49447 Add bucket api tests for create and update endpoints
For MB-49447 we need to rewrite the bucket api to use validator.erl but
this would be a large change with a risk of missing some functionality
without thorough testing. This change adds api tests to cover all
bucket create/update parameters prior to the rewrite, in order to
ensure that there is minimal unwanted change in behaviour of the
endpoint after the rewrite.

Some cases are not covered by these tests, including:
- Combinations of multiple unrelated parameters. To cover all of these
  cases would take an unreasonable amount of time, so instead only
  relevant combinations are explicitly tested, and other parameters are
  given random valid values in order to have some coverage of
  combinations of unrelated parameters.
- Thorough testing of pitrGranularity and pitrMaxHistoryAge. These give
  hard to predict errors when combined so some cases have been ignored.
  These cases could be included, they would just require precision in
  the error prediction.
- Attempting to set ram quota less than current usage. This would
  require filling a bucket above the minimum ram quota (100MiB).

Change-Id: Ia63d9a48089d6fcb0ccb4df856ca3719ee7b4df0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/182681
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>