MB-24487: Add saml authentication
IDP = Identity Provider
SP = Service Provider
SSO = Single Sign On

How it works:

UI is supposed to call

/sso/<sso_name>/auth

in order to start the authentication. That endpoint creates a saml
request and sends to browser as a redirect to IDP web page.

Then user performs authentication on the IDP webside, and if
successful, IDP redirects user back to our endpoint:

/sso/<sso_name>/samlConsume

which receives the assertion

Then ns_server checks the assertion and if it is valid the user
is considered authenticated.

Identity is extracted from the assertion.
It can be taken from NameID element (in Subject), or from
any user's attribute (for example uid, or mail).

/sso/<sso_name>/samlMetadata is called by the IDP when it needs
collect information about the SP.

Change-Id: I1760db0694fb041f6a4f35108a57d556a5e7c5ff
Reviewed-on: https://review.couchbase.org/c/ns_server/+/183995
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>