api_tests: Ensure rebalance is complete before running tests
cluster_run_lib doesn't wait for rebalance to complete during cluster
connect, so api_tests start in a rebalance for the first test.

Adds a do_wait_for_rebalance arg to cluster_run_lib connect which is
False by default. When this arg is true, we wait for rebalance to
complete with a 10min timeout.

Change-Id: If7f9e1b546ebf626deb740caaa842c8a40e0ea30
Reviewed-on: https://review.couchbase.org/c/ns_server/+/184576
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>