MB-54652 do quorum read while fetching snapshot for rebalance
This eliminates a race condition between setting the recovery
type and starting rebalance if both REST calls are done on
non orchestrator node.

I decided that blocking orchestrator with this call is somewhat
ok, because this call get's timeouted in quorum loss situations
and in these situations we are not capable to do automatic
failovers anyway.

We also currently do quorum read when we initiate graceful failover

This commit fixes the race conditions on fully chronicle clusters
only.

Change-Id: Iedda48264d229c17aa4700a6af0584935da7214a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/184490
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>