MB-54250: ensure config doesn't contain couchdb related keys
To remove couchdb (views), which we don't plan to support in
serverless. It was required to understand how the SDK's use this
information from the server. Originally, it was thought that the java
SDK would be better served by using an empty string for one of the
fields, to avoid an NPE (null pointer exception), but this still
indicated support for views. We are removing it, and accepting the
breakage with old SDK's. It doesn't seem clear that there's any way
around the SDK breakage, unfortunately.

Change-Id: I1b8ce9d35b18ea3565f1e51583ed913e0bebcb55
Reviewed-on: https://review.couchbase.org/c/ns_server/+/184483
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>