MB-52223 Revert "MB-46881: Fix the scope limit check"
This reverts commit 16c8bd5f0265a27338496a13fd966ccf07edfbaa.

Revert "MB-46881: Add fts num_queries_per_min limit."

This reverts commit 159e22c4222443f996cce33c99211b30ce6dffba.

This change is supposed to be reverted here:
 "MB-46881: User UUID propogation to services via cbauth and isasl"

But later we decided not to revert this one back. That's because
other services are using UUID and we need to keep it.

Two commits are reverted by this one.

Also this commit
Revert "MB-46881: Introduce UUID for local users"
commit c412f4823b5debd6969d98bd7921f2a0f7252236.

is not being reverted for the same reason.

Change-Id: If74949a1fc5799e39c56bffd529e4af3ba9254f9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/184773
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>