MB-57062: Add kv stuck rebalance alerts
Each service has a threshold of how long the service is expected to go
without any forwards progress. If no progress is seen over the time
period of the threshold, we fire an alert. We divide that time period
by the check frequency to get the interval between progress checks for
the service. This gives us an earlier alert than if we merely checked
once each time period, without changing the sensitivity.

This patch only checks for kv progress. To determine if a kv rebalance
is making progress we consider the seqno of each vbucket associated
with an in-progress move. We take the minimum seqno over nodes involved
with the move, for each vbucket move. We then define progress as any
change in the list of minimum seqnos of moves in progress. This
definition of progress counts the following cases as progress:
- A vbucket move starts (adding to the list of seqnos)
- The furthest behind replica in a move sees it's seqno increase
- A vbucket move completes
This means that we will only alert once all vbucket moves currently in
progress are stuck. If no moves are in progress then we don't check for
stuckness, since we are only concerned with when vbucket moves
themselves are stuck, rather than the rebalance being slow.

The limit can be configured with:
POST localhost:9000/settings/alerts/limits
"stuckRebalanceThresholdKV=36000"

Change-Id: If7f87188f106bcacd04b2110f6c20b5ba8ef0279
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188747
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>