MB-53292 delete unused memcached files right after rebalance if
...bucket placer is enabled. Before the rebalance was never able
to remove the bucket from the node, but with the introduction of
bucket placer it now can. So it make sense right after the KV rebalance
to contact alive KV nodes and ask them to remove the files that are
no longer used due to bucket being removed from the node by bucket
placer.

Change-Id: I3094446cc38a2b3367abf752398bfd81197fd997
Reviewed-on: https://review.couchbase.org/c/ns_server/+/185885
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>