Upload terses bucket info on memcached restart
Prior to this change if memcached restarted (e.g. due to OOM kill) we
wouldn't upload the terse bucket info to memcached. As a result SDKs
would not be able to boostrap.

This changes the terse_bucket_info_uploader to use the same mechanism
as the memcached_config_mgr to monitor memcached and restart if
memcached goes down. When the terse_bucket_info_uploader starts again
it will refresh the terse bucket info in memcached.

Change-Id: If3b6bd0cb0a283a046c75845c82d0106c9673d3e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186072
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>