Erlang diff formatter using emacs
Add a set of scripts that format changed regions in *.erl files using
emacs erlang-mode. Component scripts can be run individually to manually
format erlang files or get git diffs.

By default uses the erlang dependency in the build directory. Another
version of erlang may be used for formatting (in particular another
version of the erlang emacs tools) using the --erlang option, but this
is not recommended as formatting may differ and some versions of the
erlang emacs tools have been observed to not work.

By default uses 1 line of context around a diff to format according to
the current level of indentation. A "function" mode can be used which
will format an entire changed function too. There are sometimes
differences between these formatting options if not enough context
exists. The region (with 1 line of context) formatting mode is the
default as it provides minimial formatting changes if, say, only one
line of an existing function were modified. In the future when we change
and re-format more of the code base the function formatting mode may be
a more appropriate default.

The script is not capable of wrapping lines at 80 characters. The
default behaviour of emacs erlang mode does not wrap. It looks to be
possible to wrap when editing files in emacs, but I'm unsure how to do
that programatically.

The script has the ability to check the formatting of changes. It will
return an error if formatting is not the same as the formatting applied
by the script. This writes to, and deletes, a temporary file.

RUN:

    Requires emacs in your PATH:
    $ brew install emacs

    Format entire git diff:
    $ python3 scripts/formatter/erlang_formatter.py

    Get regions that will be formatted:
    $ python3 scripts/formatter/get_diffs_to_format.py

    Format file with emacs
    $ python3 scripts/formatter/format_erlang_diff_with_emacs.py -f src/collections.erl

    Check formatting
    $ python3 scripts/formatter/erlang_formatter.py --format-output-mode check

Change-Id: I1aec818b704eedc7e64e19390111222ab7f2de38
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186431
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>