Add support for filtering ns_test via cmake
This was previously merged:
https://review.couchbase.org/c/ns_server/+/182603

And reverted:
https://review.couchbase.org/c/ns_server/+/185985

As the filter was supplied as "" then TEST_FILTER was not set in the
cmake. This resulted in `make test` or `ninja ns_test` running no
tests rather than all of the tests. This has been corrected by removing
the quotation marks from do-test.cmake as they can be added when
specifying TEST_FILTER instead.

Change-Id: Iebc757c8560ce740d89a84ccbd6806864ec505a1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186115
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>