MB-55464: Add task_id to /sampleBuckets/install response
For each sample that has no immediate errors, respond with a task_id,
identifying the task responsible for loading that sample. To avoid
future backwards compatibility issues, we respond with a json object
for each task_id, containing the sample and bucket names, and identify
the list of tasks with a "tasks" key. For example:

{
    "tasks": [
        {
            "taskId": "90a484e3-946d-4567-9a8c-3dd234e1fe52",
            "bucket": "default",
            "sample": "travel-sample"
        }
    ]
}

Similarly, errors are identified with an errors key.

By separately identifying errors and tasks, we make the endpoint more
robust, and allow simpler future additions without breaking current
handling.

Change-Id: Icbd17c3f079dff58621c256af497cc84ecf697e5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186299
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>