MB-54926 Replacing the returned formatted error with an error tuple
Netconfig_update returns the formatted error string which is not
something we want later. Here we're changing it to return raw
information about the error and format it later when we need it.
Most of the time we'll be using the same formatting function
provided in this module, but we have the freedom to use any
other formatting function.

Change-Id: Ie5ea44110a435013067af2160dbb325de34f9eac
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186346
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>