MB-52758: improve default ciphers list
Due to the 'LUCKY13' class of vulnerabilities, it is required that we
upgrade our default list of accepted cipher suites. Specifically we
are trying to remove:

 - TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA
 - TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA
 - TLS_RSA_WITH_AES_256_CBC_SHA
 - TLS_RSA_WITH_AES_128_CBC_SHAa

But, we are also adding in more secure ciphers:

 - TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256
 - TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256
 - TLS_DHE_RSA_WITH_AES_128_GCM_SHA256
 - TLS_DHE_RSA_WITH_AES_256_GCM_SHA384

These are being added to the "high" security cipher group.

Change-Id: Ia6712c1d1e3ce0334f12cf80999e6c6aa14689a9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187156
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>