MB-53536: introduce internal_ssl_minimum_protocol
Introduced the new option "internal_ssl_minimum_protocol", which,
unlike "ssl_minimum_protocol" controls only couchbase internal
services. The most obvioust case is the erlang distribution
connections which can easily be set to a higher TLS version than the
external interfaces because we only need to maintain compatibility
with ourselves and don't need to support older tls versions.

All other use-cases for this are meant to be internal. That means we
don't need to maintain support for tlsv1.2, when all our services can
support a more aggressive option of tlsv1.3. There are features of
tlsv1.3 that make it preferable to previous versions, and this would
allow us to use it internally without messing up our existing
settings.

The camelcase version of this variable is: 'internalTlsMinVersion'.

Change-Id: I9435f6e03738c6f79180fed1e7fac411f20b58f2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186574
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>