MB-55427: Allow modification of collection if props are same
Currently we do not allow the modification of a collection if we pass
InvalidProps which have the same value. This causes an issue on the
set manifest path which may attempt to apply a manifest with the same
parameters and be unable to do so as a result.

Change-Id: If077eabbbdac5f467d5536315ea94df97726152b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186788
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>