MB-55624 Fix adding a 7.2 node to a 6.6 cluster failure
When adding a 7.2+ node a 6.6 cluster, if the node is using generated
certificates, the SAN check fails in some cases. The reason is the
certificates on the new node doesn't get refreshed and this leads to
SAN check to fail. Here we're ignoring the test if certs are
generated and cluster is older than 7.0.

Change-Id: I2ac8b77bacf76fdb59e9ef556f7aa6a3fbc0faab
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186927
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>