MB-55395 Improve Metrics Definitions: Cluster Manager
This adds a file containing a JSON object with the stats returned by
ns_server's _prometheusMetrics and _prometheusMetricsHigh endpoints. The
JSON is intended for use by the docs team to create applicable
documentation. In a future release the JSON will be used to provide
openMetrics "metadata" info for returned stats.

The document on which this work is based is titled "Complete Metrics
Documentation in 7.x".

As part of installation the ns_server metrics_metadata.json file should
be installed into install/etc/couchbase/cm/

Change-Id: Icf49452fd077a9552ce61246f242398e5df9d157
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186928
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>