MB-55736: Set history on collections during upgrade to 7.2.0
We should set the history field for collections on the upgrade to 7.2.0
if it is appropriate to do so (i.e. the
historyRetentionCollectionDefault is set to true). This makes the user
experience for an upgraded bucket the same as a newly created bucket
when it comes to enabling history, i.e. the customer only has to enable
history at the Bucket level, and does not have to enable it for each
collection.

Change-Id: I899577cde099196b3faf465a0087f3a2df24e24c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187457
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>