Remove api_test from ctest
Instead add it as a custom target so we can run it separately in the
api_test job which will be enabled when this patch is submitted.
This stops the unit-simple-test job from failing, until we can get a
proper fix for that.

Any patches that are tested in jenkins *without* this patch will still
run api-test through ctest as part of the ns-server-ns-test job.
These patches will receive a -1 from the api-test job, as it will fail
due to the missing make target.

Any patches that are tested in jenkins *with* this patch will no longer
run api-test as part of the ns-server-ns-test job, but will instead run
the tests in the api-test job, which will contribute to the Build Bot
Well-Formed vote, in the same way as before.

Change-Id: Ic85233a4f0f278fd1918e1b986f806880505b5fb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187651
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>