MB-55784 Use httpc persistent connections
The erlang httpc implementation has a short-coming where specifying
socket_ops to httpc:request will lead to the connection not being
persistent. This means a new connection will be used for each request
rather than reusing a connection.

To avoid this we check the ipfamily that httpc is using and if it is
different then what is expected we use httpc:set_options to set the
ipfamily thus avoiding specifying socket_opt to httpc:request.

Change-Id: Ibe120d4c93bce6f847e896b670d116cf65e1802c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187735
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>