MB-55464: Update tasks version when global tasks list changes
To maintain consistency with the prior behaviour, we should update
the version retrieved with ns_doctor:get_tasks_version when the list of
global_tasks:get_default_tasks changes. These are the tasks that we
will include by default in the /pools/default/tasks endpoint, so
the version should change whenever a new task is added.
We don't update the version when an exisiting task changes it's status.

Change-Id: Ie5ac5a159f00f2e882baa8299e95f9eaa6d0e9fa
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187818
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>