MB-55866 Add support for 'http(s)://' when importing samples
The initial plan to load samples in Elixir was to make use of the
existing infrastructure from 'cbbackupmgr'.

This posed some issues with authentication, and was all-round a
sub-optimal solution.

Samples will now be loaded using a pre-signed link, which will be passed
in place of the staging/region.

NOTE: Removed support for the 's3://' prefix, which has also been
removed from 'cbimport'.

Change-Id: Icb4379aceb6d8ce8bd502fed17c724efa5e072f8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187854
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: James Lee <james.lee@couchbase.com>