MB-55856 Not able to add 7.2.0 node into 7.0.4 cluster
When adding a 7.2+ node to a 7.0.x or older cluster, if the node
is using generated certificates, the SAN check fails in some cases.
The reason is the certificates on the new node doesn't get refreshed
due to the fact the PKey for CA is being sent only when the
encryption is on and in other cases we don't. This leads to SAN
check to fail. Here we're ignoring the test if certs are
generated and cluster is older than 7.1.

Change-Id: I7d656725d05ad5ce21e1a8110e5476aea425fa43
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187864
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>