Allow partial and multiple task updates
For sample_loader_tasks:terminating and when terminating other
gen_servers, we wish to mark all currently running tasks as failed.
To do this efficiently, we should allow updating multiple tasks in one
call, i.e. with global_tasks:update_tasks/1 which takes a list of
tasks. The reason that this is more efficient is that we can update the
tasks list just once with one chronicle transaction, whereas originally
we would perform multiple chronicle transactions unnecessarily.

This also now allows partial task updates, by providing task_id(s) and
task update function to update_task/2 or update_tasks/2 which will give
an error when attempted with a task that doesn't already exist. As part
of this change, we require that mutable fields (except status, and
fields internal to global_tasks.erl) are only found in extras.

The 'bucket' and 'bucket_uuid' fields are mutable, so they now go in
extras. Since we can determine if a task's bucket is relevant still by
checking the bucket_uuid, there is no need to remove tasks which have
a bucket_uuid that no longer matches the bucket, as the task will still
be expired by time anyway. That form of task cleanup is removed.

Change-Id: Iba83fcbb52dc374b8b50f492e0f362c427cd9a54
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187919
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>