MB-53134 bucket_placer to attempt to place buckets on delta recovery
...nodes if buckets were previously residing on those nodes

We use {node, N, failover_vbuckets} key to determine on which nodes
the bucket was residing before the failover and then give bucket_placer
a hint to prioritize those nodes when placing the bucket.

It might happen that bucket_placer won't be able to place the bucket
on the exact set of nodes on which it was residing before the failover.
Then the delta recovery of the bucket will be impossible.

Change-Id: I92e7af534b636dc2edb1d300d810ccdcef212807
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188047
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>