api_tests: Only remove cluster directories after killing nodes
Since we don't need cluster data (logs, data, etc) when the tests pass,
the temp cluster directories should be removed.
If we removed the cluster directories before killing the nodes, then we
could have this fail with the following error:

OSError: [Errno 66] Directory not empty: 'n_0'

This is because the cluster may recreate files if it is still alive,
which can prevent directory deletion as the directory becomes non-empty
unexpectedly.

Change-Id: I37b35154fc072fe07b595935b7171cb08ddd6598
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188025
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>